rename index.spec to index since it doesn't contain any test.
